dags/requirements.txt
astronomer-cosmos==1.0.3
dbt-fabric==1.5.0